+++ cedar-policy-core/src/ast/entity.rs
@@ -28,38 +28,72 @@ use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, TryFromInto};
 use smol_str::SmolStr;
 use std::collections::{BTreeMap, HashMap, HashSet};
+use std::str::FromStr;
 use thiserror::Error;
 
-/// We support two types of entities. The first is a nominal type (e.g., User, Action)
-/// and the second is an unspecified type, which is used (internally) to represent cases
-/// where the input request does not provide a principal, action, and/or resource.
+/// The entity type that Actions must have
+pub static ACTION_ENTITY_TYPE: &str = "Action";
+
+/// Entity type names are just [`Name`]s, but we have some operations on them specific to entity types.
 #[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone, Hash, PartialOrd, Ord)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
-pub enum EntityType {
-    /// Concrete nominal type
-    Specified(Name),
-    /// Unspecified
-    Unspecified,
-}
+#[serde(transparent)]
+pub struct EntityType(Name);
 
 impl EntityType {
-    /// Is this an Action entity type
+    /// Is this an Action entity type?
+    /// Returns true when an entity type is an action entity type. This compares the
+    /// base name for the type, so this will return true for any entity type named
+    /// `Action` regardless of namespaces.
     pub fn is_action(&self) -> bool {
-        match self {
-            Self::Specified(name) => name.basename() == &Id::new_unchecked("Action"),
-            Self::Unspecified => false,
-        }
+        self.0.basename() == &Id::new_unchecked(ACTION_ENTITY_TYPE)
+    }
+
+    /// The name of this entity type
+    pub fn name(&self) -> &Name {
+        &self.0
+    }
+
+    /// The source location of this entity type
+    pub fn loc(&self) -> Option<&Loc> {
+        self.0.loc()
+    }
+
+    /// Calls [`Name::prefix_namespace_if_unqualified`] on the underlying [`Name`]
+    pub fn prefix_namespace_if_unqualified(&self, namespace: Option<&Name>) -> Self {
+        Self(self.0.prefix_namespace_if_unqualified(namespace))
+    }
+
+    /// Wraps [`Name::from_normalized_str`]
+    pub fn from_normalized_str(src: &str) -> Result<Self, ParseErrors> {
+        Name::from_normalized_str(src).map(Self)
     }
 }
 
-// Note: the characters '<' and '>' are not allowed in `Name`s, so the display for
-// `Unspecified` never conflicts with `Specified(name)`.
+impl From<Name> for EntityType {
+    fn from(n: Name) -> Self {
+        Self(n)
+    }
+}
+
+impl AsRef<Name> for EntityType {
+    fn as_ref(&self) -> &Name {
+        &self.0
+    }
+}
+
+impl FromStr for EntityType {
+    type Err = <Name as FromStr>::Err;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let name: Name = s.parse()?;
+        Ok(Self(name))
+    }
+}
+
 impl std::fmt::Display for EntityType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Unspecified => write!(f, "<Unspecified>"),
-            Self::Specified(name) => write!(f, "{}", name),
-        }
+        write!(f, "{}", self.0)
     }
 }
 
@@ -133,7 +167,7 @@ impl EntityUID {
     pub(crate) fn test_entity_type() -> EntityType {
         let name = Name::parse_unqualified_name("test_entity_type")
             .expect("test_entity_type should be a valid identifier");
-        EntityType::Specified(name)
+        EntityType(name)
     }
     // by default, Coverlay does not track coverage for lines after a line
     // containing #[cfg(test)].
@@ -144,7 +178,7 @@ impl EntityUID {
     /// Create an `EntityUID` with the given (unqualified) typename, and the given string as its EID.
     pub fn with_eid_and_type(typename: &str, eid: &str) -> Result<Self, ParseErrors> {
         Ok(Self {
-            ty: EntityType::Specified(Name::parse_unqualified_name(typename)?),
+            ty: EntityType(Name::parse_unqualified_name(typename)?),
             eid: Eid(eid.into()),
             loc: None,
         })
@@ -161,22 +195,9 @@ impl EntityUID {
         self.loc.as_ref()
     }
 
-    /// Create a nominally-typed `EntityUID` with the given typename and EID
-    pub fn from_components(name: Name, eid: Eid, loc: Option<Loc>) -> Self {
-        Self {
-            ty: EntityType::Specified(name),
-            eid,
-            loc,
-        }
-    }
-
-    /// Create an unspecified `EntityUID` with the given EID
-    pub fn unspecified_from_eid(eid: Eid) -> Self {
-        Self {
-            ty: EntityType::Unspecified,
-            eid,
-            loc: None,
-        }
+    /// Create an [`EntityUID`] with the given typename and [`Eid`]
+    pub fn from_components(ty: EntityType, eid: Eid, loc: Option<Loc>) -> Self {
+        Self { ty, eid, loc }
     }
 
     /// Get the type component.
@@ -588,14 +609,16 @@ mod test {
     fn test_euid_equality() {
         let e1 = EntityUID::with_eid("foo");
         let e2 = EntityUID::from_components(
-            Name::parse_unqualified_name("test_entity_type").expect("should be a valid identifier"),
+            Name::parse_unqualified_name("test_entity_type")
+                .expect("should be a valid identifier")
+                .into(),
             Eid("foo".into()),
             None,
         );
-        let e3 = EntityUID::unspecified_from_eid(Eid("foo".into()));
-        let e4 = EntityUID::unspecified_from_eid(Eid("bar".into()));
-        let e5 = EntityUID::from_components(
-            Name::parse_unqualified_name("Unspecified").expect("should be a valid identifier"),
+        let e3 = EntityUID::from_components(
+            Name::parse_unqualified_name("Unspecified")
+                .expect("should be a valid identifier")
+                .into(),
             Eid("foo".into()),
             None,
         );
@@ -603,21 +626,12 @@ mod test {
         // an EUID is equal to itself
         assert_eq!(e1, e1);
         assert_eq!(e2, e2);
-        assert_eq!(e3, e3);
 
         // constructing with `with_euid` or `from_components` is the same
         assert_eq!(e1, e2);
 
         // other pairs are not equal
         assert!(e1 != e3);
-        assert!(e1 != e4);
-        assert!(e1 != e5);
-        assert!(e3 != e4);
-        assert!(e3 != e5);
-        assert!(e4 != e5);
-
-        // e3 and e5 are displayed differently
-        assert!(format!("{e3}") != format!("{e5}"));
     }
 
     #[test]
@@ -630,5 +644,10 @@ mod test {
         assert!(!euid.is_action());
         let euid = EntityUID::from_str("Action::Foo::\"view\"").unwrap();
         assert!(!euid.is_action());
+    }
+
+    #[test]
+    fn action_type_is_valid_id() {
+        assert!(Id::from_normalized_str(ACTION_ENTITY_TYPE).is_ok());
     }
 }