+++ cedar-policy-core/src/entities.rs
@@ -1949,12 +1949,12 @@ mod schema_based_parsing_tests {
             basename: &'a Id,
         ) -> Box<dyn Iterator<Item = EntityType> + 'a> {
             match basename.as_ref() {
-                "Employee" => Box::new(std::iter::once(EntityType::Specified(
-                    Name::unqualified_name(basename.clone()),
-                ))),
-                "Action" => Box::new(std::iter::once(EntityType::Specified(
-                    Name::unqualified_name(basename.clone()),
-                ))),
+                "Employee" => Box::new(std::iter::once(EntityType::from(Name::unqualified_name(
+                    basename.clone(),
+                )))),
+                "Action" => Box::new(std::iter::once(EntityType::from(Name::unqualified_name(
+                    basename.clone(),
+                )))),
                 _ => Box::new(std::iter::empty()),
             }
         }
@@ -1967,7 +1967,7 @@ mod schema_based_parsing_tests {
     struct MockEmployeeDescription;
     impl EntityTypeDescription for MockEmployeeDescription {
         fn entity_type(&self) -> EntityType {
-            EntityType::Specified(Name::parse_unqualified_name("Employee").expect("valid"))
+            EntityType::from(Name::parse_unqualified_name("Employee").expect("valid"))
         }
 
         fn attr_type(&self, attr: &str) -> Option<SchemaType> {
@@ -1975,7 +1975,7 @@ mod schema_based_parsing_tests {
                 ty: self.entity_type(),
             };
             let hr_ty = || SchemaType::Entity {
-                ty: EntityType::Specified(Name::parse_unqualified_name("HR").expect("valid")),
+                ty: EntityType::from(Name::parse_unqualified_name("HR").expect("valid")),
             };
             match attr {
                 "isFullTime" => Some(SchemaType::Bool),
@@ -3035,7 +3035,7 @@ mod schema_based_parsing_tests {
                 basename: &'a Id,
             ) -> Box<dyn Iterator<Item = EntityType> + 'a> {
                 match basename.as_ref() {
-                    "Employee" => Box::new(std::iter::once(EntityType::Specified(
+                    "Employee" => Box::new(std::iter::once(EntityType::from(
                         Name::from_str("XYZCorp::Employee").expect("valid name"),
                     ))),
                     _ => Box::new(std::iter::empty()),
@@ -3049,7 +3049,7 @@ mod schema_based_parsing_tests {
         struct MockEmployeeDescription;
         impl EntityTypeDescription for MockEmployeeDescription {
             fn entity_type(&self) -> EntityType {
-                EntityType::Specified("XYZCorp::Employee".parse().expect("valid"))
+                "XYZCorp::Employee".parse().expect("valid")
             }
 
             fn attr_type(&self, attr: &str) -> Option<SchemaType> {