+++ cedar-policy/src/api.rs
@@ -110,7 +110,6 @@ pub mod entities {
 }
 
 /// Entity datatype
-// INVARIANT(UidOfEntityNotUnspecified): The `EntityUid` of an `Entity` cannot be unspecified
 #[repr(transparent)]
 #[derive(Debug, Clone, PartialEq, Eq, RefCast)]
 pub struct Entity(ast::Entity);
@@ -197,7 +196,6 @@ impl Entity {
     /// assert_eq!(alice.uid(), euid);
     /// ```
     pub fn uid(&self) -> EntityUid {
-        // INVARIANT: By invariant on self and `EntityUid`: Our Uid can't be unspecified
         EntityUid::new(self.0.uid().clone())
     }
 
@@ -697,7 +695,6 @@ impl Entities {
             Dereference::Residual(_) | Dereference::NoSuchEntity => None,
             Dereference::Data(e) => Some(e),
         }?;
-        // Invariant: No way to write down the unspecified EntityUid, so no way to have ancestors that are unspecified
         Some(entity.ancestors().map(EntityUid::ref_cast))
     }
 
@@ -760,7 +757,7 @@ impl Authorizer {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
+    /// # let request: Request = Request::new(p, a, r, c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -814,7 +811,7 @@ impl Authorizer {
     ///
     /// let c = Context::empty();
     ///
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
+    /// let request: Request = Request::new(p, a, r, c, None).unwrap();
     ///
     /// // create a policy
     /// let s = r#"
@@ -1020,7 +1017,7 @@ impl Diagnostics {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
+    /// # let request: Request = Request::new(p, a, r, c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -1080,7 +1077,7 @@ impl Diagnostics {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
+    /// # let request: Request = Request::new(p, a, r, c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -2503,7 +2500,6 @@ impl Policy {
     /// Get the scope constraint on this policy's action
     pub fn action_constraint(&self) -> ActionConstraint {
         // Clone the data from Core to be consistant with the other constraints
-        // INVARIANT: all of the EntityUids come from a policy, which must have Concrete EntityUids
         match self.ast.template().action_constraint() {
             ast::ActionConstraint::Any => ActionConstraint::Any,
             ast::ActionConstraint::In(ids) => ActionConstraint::In(
@@ -2551,7 +2547,6 @@ impl Policy {
         slot: ast::SlotId,
     ) -> &'a EntityUid {
         match r {
-            // INVARIANT: this comes from policy source, so must be concrete
             ast::EntityReference::EUID(euid) => EntityUid::ref_cast(euid),
             // PANIC SAFETY: This `unwrap` here is safe due the invariant (values total map) on policies.
             #[allow(clippy::unwrap_used)]
@@ -3066,21 +3061,10 @@ impl<S> RequestBuilder<S> {
     ///
     /// Note that you can create the `EntityUid` using `.parse()` on any
     /// string (via the `FromStr` implementation for `EntityUid`).
-    ///
-    /// Here, passing `None` for `principal` indicates that `principal` does
-    /// not contribute to authorization decisions (e.g., because it is not
-    /// used in your policies).
-    /// This is different than Unknown for partial-evaluation purposes.
     #[must_use]
-    pub fn principal(self, principal: Option<EntityUid>) -> Self {
+    pub fn principal(self, principal: EntityUid) -> Self {
         Self {
-            principal: match principal {
-                Some(p) => ast::EntityUIDEntry::concrete(p.into(), None),
-                None => ast::EntityUIDEntry::concrete(
-                    ast::EntityUID::unspecified_from_eid(ast::Eid::new("principal")),
-                    None,
-                ),
-            },
+            principal: ast::EntityUIDEntry::known(principal.into(), None),
             ..self
         }
     }
@@ -3089,21 +3073,10 @@ impl<S> RequestBuilder<S> {
     ///
     /// Note that you can create the `EntityUid` using `.parse()` on any
     /// string (via the `FromStr` implementation for `EntityUid`).
-    ///
-    /// Here, passing `None` for `action` indicates that `action` does
-    /// not contribute to authorization decisions (e.g., because it is not
-    /// used in your policies).
-    /// This is different than Unknown for partial-evaluation purposes.
     #[must_use]
-    pub fn action(self, action: Option<EntityUid>) -> Self {
+    pub fn action(self, action: EntityUid) -> Self {
         Self {
-            action: match action {
-                Some(a) => ast::EntityUIDEntry::concrete(a.into(), None),
-                None => ast::EntityUIDEntry::concrete(
-                    ast::EntityUID::unspecified_from_eid(ast::Eid::new("action")),
-                    None,
-                ),
-            },
+            action: ast::EntityUIDEntry::known(action.into(), None),
             ..self
         }
     }
@@ -3112,21 +3085,10 @@ impl<S> RequestBuilder<S> {
     ///
     /// Note that you can create the `EntityUid` using `.parse()` on any
     /// string (via the `FromStr` implementation for `EntityUid`).
-    ///
-    /// Here, passing `None` for `resource` indicates that `resource` does
-    /// not contribute to authorization decisions (e.g., because it is not
-    /// used in your policies).
-    /// This is different than Unknown for partial-evaluation purposes.
     #[must_use]
-    pub fn resource(self, resource: Option<EntityUid>) -> Self {
+    pub fn resource(self, resource: EntityUid) -> Self {
         Self {
-            resource: match resource {
-                Some(r) => ast::EntityUIDEntry::concrete(r.into(), None),
-                None => ast::EntityUIDEntry::concrete(
-                    ast::EntityUID::unspecified_from_eid(ast::Eid::new("resource")),
-                    None,
-                ),
-            },
+            resource: ast::EntityUIDEntry::known(resource.into(), None),
             ..self
         }
     }
@@ -3214,28 +3176,16 @@ impl Request {
     /// If `schema` is present, this constructor will validate that the
     /// `Request` complies with the given `schema`.
     pub fn new(
-        principal: Option<EntityUid>,
-        action: Option<EntityUid>,
-        resource: Option<EntityUid>,
+        principal: EntityUid,
+        action: EntityUid,
+        resource: EntityUid,
         context: Context,
         schema: Option<&Schema>,
     ) -> Result<Self, RequestValidationError> {
-        let p = principal.map_or_else(
-            || ast::EntityUID::unspecified_from_eid(ast::Eid::new("principal")),
-            EntityUid::into,
-        );
-        let a = action.map_or_else(
-            || ast::EntityUID::unspecified_from_eid(ast::Eid::new("action")),
-            EntityUid::into,
-        );
-        let r = resource.map_or_else(
-            || ast::EntityUID::unspecified_from_eid(ast::Eid::new("resource")),
-            EntityUid::into,
-        );
         Ok(Self(ast::Request::new(
-            (p, None),
-            (a, None),
-            (r, None),
+            (principal.into(), None),
+            (action.into(), None),
+            (resource.into(), None),
             context.0,
             schema.map(|schema| &schema.0),
             Extensions::all_available(),
@@ -3243,43 +3193,28 @@ impl Request {
     }
 
     /// Get the principal component of the request. Returns `None` if the principal is
-    /// "unspecified" (i.e., constructed by passing `None` into the constructor) or
     /// "unknown" (i.e., constructed using the partial evaluation APIs).
     pub fn principal(&self) -> Option<&EntityUid> {
         match self.0.principal() {
-            ast::EntityUIDEntry::Known { euid, .. } => match euid.entity_type() {
-                // INVARIANT: we ensure Concrete-ness here
-                ast::EntityType::Specified(_) => Some(EntityUid::ref_cast(euid.as_ref())),
-                ast::EntityType::Unspecified => None,
-            },
+            ast::EntityUIDEntry::Known { euid, .. } => Some(EntityUid::ref_cast(euid.as_ref())),
             ast::EntityUIDEntry::Unknown { .. } => None,
         }
     }
 
     /// Get the action component of the request. Returns `None` if the action is
-    /// "unspecified" (i.e., constructed by passing `None` into the constructor) or
     /// "unknown" (i.e., constructed using the partial evaluation APIs).
     pub fn action(&self) -> Option<&EntityUid> {
         match self.0.action() {
-            ast::EntityUIDEntry::Known { euid, .. } => match euid.entity_type() {
-                // INVARIANT: we ensure Concrete-ness here
-                ast::EntityType::Specified(_) => Some(EntityUid::ref_cast(euid.as_ref())),
-                ast::EntityType::Unspecified => None,
-            },
+            ast::EntityUIDEntry::Known { euid, .. } => Some(EntityUid::ref_cast(euid.as_ref())),
             ast::EntityUIDEntry::Unknown { .. } => None,
         }
     }
 
     /// Get the resource component of the request. Returns `None` if the resource is
-    /// "unspecified" (i.e., constructed by passing `None` into the constructor) or
     /// "unknown" (i.e., constructed using the partial evaluation APIs).
     pub fn resource(&self) -> Option<&EntityUid> {
         match self.0.resource() {
-            ast::EntityUIDEntry::Known { euid, .. } => match euid.entity_type() {
-                // INVARIANT: we ensure Concrete-ness here
-                ast::EntityType::Specified(_) => Some(EntityUid::ref_cast(euid.as_ref())),
-                ast::EntityType::Unspecified => None,
-            },
+            ast::EntityUIDEntry::Known { euid, .. } => Some(EntityUid::ref_cast(euid.as_ref())),
             ast::EntityUIDEntry::Unknown { .. } => None,
         }
     }
@@ -3314,7 +3249,7 @@ impl Context {
     /// # let p = EntityUid::from_str(r#"User::"alice""#).unwrap();
     /// # let a = EntityUid::from_str(r#"Action::"view""#).unwrap();
     /// # let r = EntityUid::from_str(r#"Album::"trip""#).unwrap();
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
+    /// # let request: Request = Request::new(p, a, r, context, None).unwrap();
     /// ```
     pub fn from_pairs(
         pairs: impl IntoIterator<Item = (String, RestrictedExpression)>,
@@ -3352,7 +3287,7 @@ impl Context {
     /// # let p = EntityUid::from_str(r#"User::"alice""#).unwrap();
     /// # let a = EntityUid::from_str(r#"Action::"view""#).unwrap();
     /// # let r = EntityUid::from_str(r#"Album::"trip""#).unwrap();
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
+    /// # let request: Request = Request::new(p, a, r, context, None).unwrap();
     /// ```
     pub fn from_json_str(
         json: &str,
@@ -3416,7 +3351,7 @@ impl Context {
     /// let context = Context::from_json_value(data, Some((&schema, &action))).unwrap();
     /// # let p = EntityUid::from_str(r#"User::"alice""#).unwrap();
     /// # let r = EntityUid::from_str(r#"Album::"trip""#).unwrap();
-    /// # let request: Request = Request::new(Some(p), Some(action), Some(r), context, Some(&schema)).unwrap();
+    /// # let request: Request = Request::new(p, action, r, context, Some(&schema)).unwrap();
     /// ```
     pub fn from_json_value(
         json: serde_json::Value,
@@ -3462,7 +3397,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
+    /// # let request: Request = Request::new(p, a, r, context, None).unwrap();
     /// ```
     pub fn from_json_file(
         json: impl std::io::Read,